+++ cli/src/tasks/copy.ts
@@ -23,9 +23,12 @@ import { getPlugins } from '../plugin';
 import { allSerial } from '../util/promise';
 import { copyWeb } from '../web/copy';
 
+import { inlineSourceMaps } from './sourcemaps';
+
 export async function copyCommand(
   config: Config,
   selectedPlatformName: string,
+  inline = false,
 ): Promise<void> {
   if (selectedPlatformName && !(await isValidPlatform(selectedPlatformName))) {
     const platformDir = resolvePlatform(config, selectedPlatformName);
@@ -43,7 +46,7 @@ export async function copyCommand(
     const platforms = await selectPlatforms(config, selectedPlatformName);
     try {
       await allSerial(
-        platforms.map(platformName => () => copy(config, platformName)),
+        platforms.map(platformName => () => copy(config, platformName, inline)),
       );
     } catch (e) {
       if (isFatal(e)) {
@@ -58,6 +61,7 @@ export async function copyCommand(
 export async function copy(
   config: Config,
   platformName: string,
+  inline = false,
 ): Promise<void> {
   await runTask(c.success(c.strong(`copy ${platformName}`)), async () => {
     const result = await checkWebDir(config);
@@ -82,6 +86,14 @@ export async function copy(
       usesCapacitorPortals = true;
     }
 
+    let usesLiveUpdates = false;
+    if (
+      allPlugins.filter(plugin => plugin.id === '@capacitor/live-updates')
+        .length > 0
+    ) {
+      usesLiveUpdates = true;
+    }
+
     if (platformName === config.ios.name) {
       if (usesCapacitorPortals) {
         await copyFederatedWebDirs(config, await config.ios.webDirAbs);
@@ -91,6 +103,9 @@ export async function copy(
           await config.ios.webDirAbs,
           config.app.webDirAbs,
         );
+      }
+      if (usesLiveUpdates) {
+        await copySecureLiveUpdatesKey(config, config.ios.nativeTargetDirAbs);
       }
       await copyCapacitorConfig(config, config.ios.nativeTargetDirAbs);
       const cordovaPlugins = await getCordovaPlugins(config, platformName);
@@ -104,6 +119,9 @@ export async function copy(
           config.android.webDirAbs,
           config.app.webDirAbs,
         );
+      }
+      if (usesLiveUpdates) {
+        await copySecureLiveUpdatesKey(config, config.android.assetsDirAbs);
       }
       await copyCapacitorConfig(config, config.android.assetsDirAbs);
       const cordovaPlugins = await getCordovaPlugins(config, platformName);
@@ -119,6 +137,9 @@ export async function copy(
       }
     } else {
       throw `Platform ${platformName} is not valid.`;
+    }
+    if (inline) {
+      await inlineSourceMaps(config, platformName);
     }
   });
 
@@ -205,5 +226,36 @@ function isPortal(config: any): config is Portal {
   return (
     (config as Portal).webDir !== undefined &&
     (config as Portal).name !== undefined
+  );
+}
+
+async function copySecureLiveUpdatesKey(config: Config, nativeAbsDir: string) {
+  if (!config.app.extConfig?.plugins?.LiveUpdates?.key) {
+    return;
+  }
+
+  const secureLiveUpdatesKeyFile = config.app.extConfig.plugins.LiveUpdates.key;
+  const keyAbsFromPath = join(config.app.rootDir, secureLiveUpdatesKeyFile);
+  const keyAbsToPath = join(nativeAbsDir, basename(keyAbsFromPath));
+  const keyRelToDir = relative(config.app.rootDir, nativeAbsDir);
+
+  if (!(await pathExists(keyAbsFromPath))) {
+    logger.warn(
+      `Cannot copy Secure Live Updates signature file from ${c.strong(
+        keyAbsFromPath,
+      )} to ${keyRelToDir}\n` +
+        `Signature file does not exist at specified key path.`,
+    );
+
+    return;
+  }
+
+  await runTask(
+    `Copying Secure Live Updates key from ${c.strong(
+      secureLiveUpdatesKeyFile,
+    )} to ${keyRelToDir}`,
+    async () => {
+      return fsCopy(keyAbsFromPath, keyAbsToPath);
+    },
   );
 }